💄 style(readme): Add text-alignment and word-wrap to p tags
This commit adds `text-align: justify` and `word-wrap: break-word` CSS properties to the `p` tag in the Markdown component and `GlobalStyle`. It also updates dependencies: `leva` from `0.9.34` to `0.9`, `remark-gfm` from `3.0.1` to `3`, and adds `pangu` and its types as new dependencies. Lastly, it uses `pangu.spacing(children)` to add proper space between Chinese characters and English words in the rendered markdown content.