💄 style(tokens): Update color palette and gray scale colors
This commit updates the tokens for colorTextTertiary, colorTextQuaternary, colorBorder, and colorBorderSecondary to use different values in generateColorPalette.ts. It also changes the light and lightA scales of the gray color in colors.ts. Finally, it modifies the ActionIcon style file by changing the value of colorTextQuaternary to colorTextTertiary and updating the value of background-color.