💄 style(syntaxhighlighter): Update variable name in codeToHtml
This commit updates the `isDarkTheme` variable name to `isDarkMode` in `codeToHtml()` function inside `SyntaxHighlighter` component. This change was made to align with the naming convention used throughout the project.

It also adds a new file `src/Highlighter/theme.ts` which contains the color palette for different themes based on `isDarkMode` parameter passed. The colors are used in the semantic highlighting of the code.