♻️ refactor(components): refactor Footer and Toc components
This commit refactors the code in the Footer and Toc components to change the container divs to HTML5 semantic elements for better accessibility. The Footer component container div is changed to footer, while Toc component's container div is changed to nav. In addition, the layout div in the Docs page is removed and replaced with a main element for better semantics.