✨ feat: Add "extra" prop to FormGroup component
This commit adds an "extra" prop to the FormGroup component in both the
FormGroup.tsx and index.tsx files. The "extra" prop is of type ReactNode
and is passed down as a prop to the FormGroup component.

The changes are done to enhance the functionality of the FormGroup component by allowing the addition of extra content.