✨ feat(components): Add "color" and "fill" props to ActionIcon and Icon components
This commit introduces the "color" and "fill" props to the ActionIcon and Icon components. The "color" prop is used to specify the color of the icon, while the "fill" prop is used to specify the fill color of the icon. These props are then passed to the Icon component when rendering the icon.

Changes:
- Added "color" prop to ActionIcon component
- Added "fill" prop to ActionIcon component
- Added "color" prop to Icon component
- Added "fill" prop to Icon component