✨ feat(chatlist): Update logic for rendering chat items based on role
This commit introduces a new approach for rendering chat items in the
"ChatList" directory's "index.tsx" file. The logic has been modified to
utilize a conditional rendering approach using the "renderItem"
function. If the "renderItem" function is defined and has a corresponding
role, it will be used to render the item. Otherwise, the default "Item"
component will be used.

Changes:
- Modify logic for rendering chat items based on role in "index.tsx"
- Utilize conditional rendering approach using "renderItem" function