✨ feat: Add banner prop to Alert component
The commit adds a new prop called "banner" to the Alert component. This prop allows the Alert component to be displayed as a banner. The banner prop is used to change the appearance of the Alert component to be more prominent and visually distinct from regular Alerts. This change enhances the visual representation of the Alert component and provides more flexibility in its usage.